feat(workers): consolidate keyframeEval, timelineSnap, and projectWorker into compute.worker
Merges three separate worker isolates into a single persistent compute domain
worker to eliminate ~40 MB of idle V8 heap overhead and prevent CPU
oversubscription with Tauri's Rust threads (FFmpeg, CPAL, wgpu).

compute.worker.ts handles:
  - Cubic Bézier root solving for keyframe interpolation
  - 1D interval-tree snapping and ripple displacement
  - Project JSON serialisation and patch diffing

Removes: keyframeEval.worker.ts, timelineSnap.worker.ts, projectWorker.worker.ts
