--- src/workers/timelineSnap.worker.ts
@@ -1,257 +1,0 @@
-/**
- * TimelineSnapWorker — Off-Thread 1D Interval Tree, Magnetic Snapping & Ripple Math
- *
- * Maintains a pre-sorted 1D index of timeline edges and markers to provide:
- * • Binary-search magnetic snapping to clip start/end, playhead, markers, and zero in < 0.5ms
- * • Fast 1D interval collision detection for dragged clips
- * • Multi-track ripple delta calculations without main-thread UI hitching
- */
-
-import type {
-  TimelineSnapWorkerRequest,
-  SnapSyncMessage,
-  SnapQueryMessage,
-  RippleComputeMessage,
-  SnapResult,
-  RippleResult,
-  SnapClip,
-  SnapMarker,
-  SnapGuideType,
-  WorkerErrorResponse,
-} from "./types";
-
-interface SnapEdge {
-  time: number;
-  guideType: SnapGuideType;
-  clipId?: string;
-  trackId?: string;
-}
-
-let storedClips: SnapClip[] = [];
-let storedMarkers: SnapMarker[] = [];
-let sortedEdges: SnapEdge[] = [];
-
-function rebuildSortedEdges(): void {
-  const edges: SnapEdge[] = [];
-
-  for (const clip of storedClips) {
-    edges.push({
-      time: clip.startTime,
-      guideType: "clip-start",
-      clipId: clip.clipId,
-      trackId: clip.trackId,
-    });
-    edges.push({
-      time: clip.startTime + clip.duration,
-      guideType: "clip-end",
-      clipId: clip.clipId,
-      trackId: clip.trackId,
-    });
-  }
-
-  for (const marker of storedMarkers) {
-    edges.push({
-      time: marker.time,
-      guideType: "marker",
-    });
-  }
-
-  edges.sort((a, b) => a.time - b.time);
-  sortedEdges = edges;
-}
-
-function handleSyncState(msg: SnapSyncMessage): void {
-  storedClips = msg.clips || [];
-  storedMarkers = msg.markers || [];
-  rebuildSortedEdges();
-}
-
-function handleSnapQuery(msg: SnapQueryMessage): void {
-  const {
-    id,
-    draggedClipId,
-    proposedStartTime,
-    trackId,
-    snapEnabled,
-    snapRadiusSeconds = 0.1,
-    playheadTime,
-  } = msg;
-
-  if (!snapEnabled) {
-    const collidingClipIds = findCollisions(draggedClipId, trackId, proposedStartTime);
-    const response: SnapResult = {
-      type: "SNAP_RESULT",
-      id,
-      snappedTime: proposedStartTime,
-      snapGuides: [],
-      collidingClipIds,
-    };
-    (self as unknown as Worker).postMessage(response);
-    return;
-  }
-
-  let closestDistance = snapRadiusSeconds;
-  let snappedTime = proposedStartTime;
-  let activeGuide: { time: number; guideType: SnapGuideType } | null = null;
-
-  // 1. Timeline start (t = 0)
-  const distToZero = Math.abs(proposedStartTime);
-  if (distToZero < closestDistance) {
-    closestDistance = distToZero;
-    snappedTime = 0;
-    activeGuide = { time: 0, guideType: "clip-start" };
-  }
-
-  // 2. Playhead snap
-  if (Number.isFinite(playheadTime)) {
-    const distToPlayhead = Math.abs(proposedStartTime - playheadTime);
-    if (distToPlayhead < closestDistance) {
-      closestDistance = distToPlayhead;
-      snappedTime = playheadTime;
-      activeGuide = { time: playheadTime, guideType: "playhead" };
-    }
-  }
-
-  // 3. Binary search in sortedEdges for closest edge to proposedStartTime
-  let low = 0;
-  let high = sortedEdges.length - 1;
-
-  while (low <= high) {
-    const mid = (low + high) >> 1;
-    const edge = sortedEdges[mid];
-
-    if (edge.clipId !== draggedClipId) {
-      const dist = Math.abs(edge.time - proposedStartTime);
-      if (dist < closestDistance) {
-        closestDistance = dist;
-        snappedTime = edge.time;
-        activeGuide = { time: edge.time, guideType: edge.guideType };
-      }
-    }
-
-    if (edge.time < proposedStartTime) {
-      low = mid + 1;
-    } else {
-      high = mid - 1;
-    }
-  }
-
-  // Check neighbors around binary search landing spot
-  const checkIdx = Math.max(0, Math.min(sortedEdges.length - 1, low));
-  for (let i = Math.max(0, checkIdx - 4); i <= Math.min(sortedEdges.length - 1, checkIdx + 4); i++) {
-    const edge = sortedEdges[i];
-    if (edge.clipId === draggedClipId) continue;
-    const dist = Math.abs(edge.time - proposedStartTime);
-    if (dist < closestDistance) {
-      closestDistance = dist;
-      snappedTime = edge.time;
-      activeGuide = { time: edge.time, guideType: edge.guideType };
-    }
-  }
-
-  const collidingClipIds = findCollisions(draggedClipId, trackId, snappedTime);
-
-  const response: SnapResult = {
-    type: "SNAP_RESULT",
-    id,
-    snappedTime,
-    snapGuides: activeGuide ? [activeGuide] : [],
-    collidingClipIds,
-  };
-
-  (self as unknown as Worker).postMessage(response);
-}
-
-function findCollisions(
-  draggedClipId: string,
-  trackId: string,
-  startTime: number,
-): string[] {
-  const dragged = storedClips.find((c) => c.clipId === draggedClipId);
-  const duration = dragged?.duration ?? 5.0;
-  const endTime = startTime + duration;
-
-  const colliding: string[] = [];
-  for (const clip of storedClips) {
-    if (clip.clipId === draggedClipId || clip.trackId !== trackId) continue;
-    const clipEnd = clip.startTime + clip.duration;
-    // Overlap condition: start < clipEnd && end > clipStart
-    if (startTime < clipEnd && endTime > clip.startTime) {
-      colliding.push(clip.clipId);
-    }
-  }
-
-  return colliding;
-}
-
-function handleRippleCompute(msg: RippleComputeMessage): void {
-  const { id, anchorClipId, side, deltaSeconds, lockedTrackIds = [] } = msg;
-
-  const anchor = storedClips.find((c) => c.clipId === anchorClipId);
-  if (!anchor || deltaSeconds === 0) {
-    const response: RippleResult = {
-      type: "RIPPLE_RESULT",
-      id,
-      clipDeltas: [],
-    };
-    (self as unknown as Worker).postMessage(response);
-    return;
-  }
-
-  const lockedSet = new Set(lockedTrackIds);
-  const anchorBoundary =
-    side === "left" ? anchor.startTime : anchor.startTime + anchor.duration;
-
-  const clipDeltas: Array<{ clipId: string; deltaSeconds: number }> = [];
-
-  for (const clip of storedClips) {
-    if (clip.clipId === anchorClipId || lockedSet.has(clip.trackId)) continue;
-    if (clip.startTime >= anchorBoundary) {
-      clipDeltas.push({
-        clipId: clip.clipId,
-        deltaSeconds,
-      });
-    }
-  }
-
-  const response: RippleResult = {
-    type: "RIPPLE_RESULT",
-    id,
-    clipDeltas,
-  };
-
-  (self as unknown as Worker).postMessage(response);
-}
-
-// ─── Worker Event Listener ───────────────────────────────────────────────────
-
-self.onmessage = (event: MessageEvent<TimelineSnapWorkerRequest>) => {
-  const msg = event.data;
-  if (!msg || typeof msg !== "object") return;
-
-  try {
-    switch (msg.type) {
-      case "SYNC_STATE":
-        handleSyncState(msg);
-        break;
-      case "SNAP_QUERY":
-        handleSnapQuery(msg);
-        break;
-      case "RIPPLE_COMPUTE":
-        handleRippleCompute(msg);
-        break;
-      case "DISPOSE":
-        storedClips = [];
-        storedMarkers = [];
-        sortedEdges = [];
-        break;
-    }
-  } catch (error) {
-    const errorResponse: WorkerErrorResponse = {
-      type: "ERROR",
-      id: "id" in msg ? msg.id : undefined,
-      message: error instanceof Error ? error.message : String(error),
-    };
-    (self as unknown as Worker).postMessage(errorResponse);
-  }
-};
